requirements.txt
azure-common==1.1.26
azure-core==1.9.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
idna==2.10
isodate==0.6.0
msrest==0.6.19
oauthlib==3.1.0
pycparser==2.20
python-dateutil==2.8.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.2
